+++ appender/ConsoleAppender.ts
@@ -6,19 +6,6 @@ import ConsoleAppenderWindowSetup = require("./ConsoleAppenderWindowSetup");
 import PatternLayout = require("../layout/PatternLayout");
 
 
-// Console extension functions
-function padWithSpaces(str: string, len: number) {
-    if (str.length < len) {
-        var spaces = [];
-        var numberOfSpaces = Math.max(0, len - str.length);
-        for (var i = 0; i < numberOfSpaces; i++) {
-            spaces[i] = " ";
-        }
-        str += spaces.join("");
-    }
-    return str;
-}
-
 // PopUpAppender and InPageAppender related
 
 function setCookie(name: string, value: string, days?: number, path?: string) {
@@ -948,10 +935,10 @@ function dir(obj: any) {
     // Create the nicely formatted property list
     var propList = [];
     for (p in obj) {
-        var propNameStr = "  " + padWithSpaces(Utils.toStr(p), maxLen + 2);
+        var propNameStr = "  " + Utils.padWithSpaces(Utils.toStr(p), maxLen + 2);
         var propVal;
         try {
-            propVal = Utils.splitIntoLines(Utils.toStr(obj[p])).join(padWithSpaces(Globals.newLine, maxLen + 6));
+            propVal = Utils.splitIntoLines(Utils.toStr(obj[p])).join(Utils.padWithSpaces(Globals.newLine, maxLen + 6));
         } catch (ex) {
             propVal = "[Error obtaining property. Details: " + Utils.getExceptionMessage(ex) + "]";
         }