+++ log-util/DateUtil.ts
@@ -1,12 +1,15 @@
 ﻿
 module DateUtil {
-    export var ONE_DAY_MILLIS = 24 * 60 * 60 * 1000;
-    export var ONE_WEEK_MILLIS = 7 * ONE_DAY_MILLIS;
+    /** Number of milliseconds in a 24 hour period */
+    export var MILLIS_PER_DAY = 1000 * 60 * 60 * 24;
+    /** Number of milliseconds in a 7 x 24 hour period */
+    export var MILLIS_PER_WEEK = 7 * MILLIS_PER_DAY;
+
     export var DEFAULT_MINIMAL_DAYS_IN_FIRST_WEEK = 1;
 
-    export var monthNames = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
+    export var MONTHS = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
 
-    export var dayNames = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
+    export var DAYS_OF_WEEK = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
 
 
     export function newMidnightDate(year: number, month: number, day: number) {
@@ -16,30 +19,20 @@ module DateUtil {
     }
 
 
-    export function getDifference(base: Date, date: Date) {
-        return <number>base.getTime() - <number>date.getTime();
-    }
-
-
     export function isBefore(base: Date, date: Date) {
         return base.getTime() < date.getTime();
     }
 
 
-    export function getUtcTime(date: Date) {
-        return Date.UTC(date.getFullYear(), date.getMonth(), date.getDate(), date.getHours(), date.getMinutes(), date.getSeconds(), date.getMilliseconds());
-    }
-
-
     export function getTimeSince(base: Date, date: Date) {
-        return getUtcTime(base) - getUtcTime(date);
+        return <number>base.getTime() - <number>date.getTime();
     }
 
 
     export function getPreviousSunday(date: Date) {
         // Using midday avoids any possibility of DST messing things up
         var midday = new Date(date.getFullYear(), date.getMonth(), date.getDate(), 12, 0, 0);
-        var previousSunday = new Date(<number>midday.getTime() - date.getDay() * ONE_DAY_MILLIS);
+        var previousSunday = new Date(<number>midday.getTime() - date.getDay() * MILLIS_PER_DAY);
         return newMidnightDate(previousSunday.getFullYear(), previousSunday.getMonth(), previousSunday.getDate());
     }
 
@@ -50,7 +43,7 @@ module DateUtil {
         }
         var previousSunday = getPreviousSunday(date);
         var startOfYear = newMidnightDate(date.getFullYear(), 0, 1);
-        var weeksSinceSartOfYear = Math.floor(getTimeSince(previousSunday, startOfYear) / ONE_WEEK_MILLIS);
+        var weeksSinceSartOfYear = Math.floor(getTimeSince(previousSunday, startOfYear) / MILLIS_PER_WEEK);
         var numberOfSundays = isBefore(previousSunday, startOfYear) ? 0 : 1 + weeksSinceSartOfYear;
         var numberOfDaysInFirstWeek = 7 - startOfYear.getDay();
         var weekInYear = numberOfSundays;
@@ -68,7 +61,7 @@ module DateUtil {
         var previousSunday = getPreviousSunday(date);
         var startOfMonth = newMidnightDate(date.getFullYear(), date.getMonth(), 1);
         var numberOfSundays = isBefore(previousSunday, startOfMonth) ?
-            0 : 1 + Math.floor(getTimeSince(previousSunday, startOfMonth) / ONE_WEEK_MILLIS);
+            0 : 1 + Math.floor(getTimeSince(previousSunday, startOfMonth) / MILLIS_PER_WEEK);
         var numberOfDaysInFirstWeek = 7 - startOfMonth.getDay();
         var weekInMonth = numberOfSundays;
         if (numberOfDaysInFirstWeek >= minimalDaysInFirstWeek) {
@@ -80,7 +73,7 @@ module DateUtil {
 
     export function getDayInYear(date: Date) {
         var startOfYear = newMidnightDate(date.getFullYear(), 0, 1);
-        return 1 + Math.floor(getTimeSince(date, startOfYear) / ONE_DAY_MILLIS);
+        return 1 + Math.floor(getTimeSince(date, startOfYear) / MILLIS_PER_DAY);
     }
 
 }