+++ log-util/Timer.js
@@ -7,6 +7,11 @@ var Timer;
     Timer.PERFORMANCE_MARK_START_SUFFIX = "-start";
     Timer.PERFORMANCE_MARK_END_SUFFIX = "-end";
     Timer.PERFORMANCE_MARK_MEASURE_SUFFIX = "-duration";
+    /** Create a timer using new Date()
+     * @param name the new timer's name
+     * @param level the log level of the timer
+     * @return a new Log4Ts.Timer instance
+     */
     function newDateInst(name, level) {
         if (level === void 0) { level = Level.INFO; }
         var startDate = new Date();
@@ -31,6 +36,11 @@ var Timer;
         return inst;
     }
     Timer.newDateInst = newDateInst;
+    /** Create a timer using window.performance.now()
+     * @param name the new timer's name
+     * @param level the log level of the timer
+     * @return a new Log4Ts.Timer instance
+     */
     function newPerformanceNowInst(name, level) {
         if (level === void 0) { level = Level.INFO; }
         var startMillis = window.performance.now();
@@ -54,6 +64,14 @@ var Timer;
         return inst;
     }
     Timer.newPerformanceNowInst = newPerformanceNowInst;
+    /** Create a timer using window.performance.mark(...)
+     * @param name the new timer's name
+     * @param level the log level of the timer
+     * @param startSuffix a suffix to append to the timer's name to make a unique start time marker, or a method which takes the timer name and returns a unique start time marker
+     * @param endSuffix a suffix to append to the timer's name to make a unique end time marker, or a method which takes the timer name and returns a unique end time marker
+     * @param measureSuffix a suffix to append to the timer's name to make a measure time marker, or a method which takes the timer name and returns a unique measure time marker
+     * @return a new Log4Ts.Timer instance
+     */
     function newPerformanceMarkInst(name, level, startSuffix, endSuffix, measureSuffix) {
         if (level === void 0) { level = Level.INFO; }
         var startName = (startSuffix != null ? (typeof startSuffix === "string" ? name + startSuffix : startSuffix(name)) : name + Timer.PERFORMANCE_MARK_START_SUFFIX);