+++ log-util/Timer.ts
@@ -8,6 +8,11 @@ module Timer {
     export var PERFORMANCE_MARK_MEASURE_SUFFIX = "-duration";
 
 
+    /** Create a timer using new Date()
+     * @param name the new timer's name
+     * @param level the log level of the timer
+     * @return a new Log4Ts.Timer instance
+     */
     export function newDateInst(name: string, level: Log4Ts.Level = Level.INFO): Log4Ts.Timer {
         var startDate = new Date();
 
@@ -33,6 +38,11 @@ module Timer {
     }
 
 
+    /** Create a timer using window.performance.now()
+     * @param name the new timer's name
+     * @param level the log level of the timer
+     * @return a new Log4Ts.Timer instance
+     */
     export function newPerformanceNowInst(name: string, level: Log4Ts.Level = Level.INFO): Log4Ts.Timer {
         var startMillis = window.performance.now();
 
@@ -57,8 +67,17 @@ module Timer {
     }
 
 
+    /** Create a timer using window.performance.mark(...)
+     * @param name the new timer's name
+     * @param level the log level of the timer
+     * @param startSuffix a suffix to append to the timer's name to make a unique start time marker, or a method which takes the timer name and returns a unique start time marker
+     * @param endSuffix a suffix to append to the timer's name to make a unique end time marker, or a method which takes the timer name and returns a unique end time marker
+     * @param measureSuffix a suffix to append to the timer's name to make a measure time marker, or a method which takes the timer name and returns a unique measure time marker
+     * @return a new Log4Ts.Timer instance
+     */
     export function newPerformanceMarkInst(name: string, level: Log4Ts.Level = Level.INFO,
             startSuffix?: string | ((name: string) => string), endSuffix?: string | ((name: string) => string), measureSuffix?: string | ((name: string) => string)): Log4Ts.Timer {
+
         var startName = (startSuffix != null ? (typeof startSuffix === "string" ? name + startSuffix : startSuffix(name)) : name + Timer.PERFORMANCE_MARK_START_SUFFIX);
         var endName = (endSuffix != null ? (typeof endSuffix === "string" ? name + endSuffix : endSuffix(name)) : name + Timer.PERFORMANCE_MARK_END_SUFFIX);
         var measureName = (measureSuffix != null ? (typeof measureSuffix === "string" ? name + measureSuffix : measureSuffix(name)) : name + Timer.PERFORMANCE_MARK_MEASURE_SUFFIX);