+++ log-util/Utils.js
@@ -31,6 +31,18 @@ var Utils;
         return str.replace(/\r\n|\r|\n/g, "\\r\\n");
     }
     Utils.escapeNewLines = escapeNewLines;
+    function padWithSpaces(str, len) {
+        if (str.length < len) {
+            var spaces = [];
+            var count = Math.max(0, len - str.length);
+            for (var i = 0; i < count; i++) {
+                spaces[i] = " ";
+            }
+            str += spaces.join("");
+        }
+        return str;
+    }
+    Utils.padWithSpaces = padWithSpaces;
     function padWithZeroes(str, len) {
         while (str.length < len) {
             str = "0" + str;